# openapi.yml
openapi: '3.0.2'
info:
  title: BitBadges API
  version: '0.1'
servers:
  - url: https://api.bitbadges.io/api/v0
paths:
  /status:
    post:
      operationId: getStatus
      summary: Get Status
      description: |
        Gets the current status details about the blockchain / indexer (gas, block height, etc).
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/iGetStatusRouteSuccessResponse'
          description: Returns the current status of the blockchain / indexer

        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
      parameters:
        - name: x-api-key
          in: header
          description: BitBadges API Key for authentication
          required: true
          schema:
            type: string
      security:
        - apiKey: [] # This specifies that an API key is required for authentication
...
  /addressList/update:
    post:
      operationId: updateAddressLists
      summary: Update address lists
      description: Updatess address lists stored by BitBadges centralized servers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAddressListsRouteRequestBody'
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/iUpdateAddressListsRouteSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
      parameters:
        - name: x-api-key
          in: header
          description: BitBadges API Key for authentication
          required: true
          schema:
            type: string
      security:
        - apiKey: [] # This specifies that an API key is required for authentication
        - userIsOwner: []
        - userSignedIn: []
  /addressList/create:
    post:
      operationId: createAddressLists
      summary: Creates address lists
      description: Creates address lists stored by BitBadges centralized servers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAddressListsRouteRequestBody'
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/iUpdateAddressListsRouteSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
      parameters:
        - name: x-api-key
          in: header
          description: BitBadges API Key for authentication
          required: true
          schema:
            type: string
      security:
        - apiKey: [] # This specifies that an API key is required for authentication
        - userIsOwner: []
        - userSignedIn: []
  /addressList:
    post:
      operationId: getAddressLists
      summary: Get address lists
      description: Gets address lists. Can be on-chain or off-chain.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetAddressListsRouteRequestBody'
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/iGetAddressListsRouteSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
      parameters:
        - name: x-api-key
          in: header
          description: BitBadges API Key for authentication
          required: true
          schema:
            type: string
      security:
        - apiKey: [] # This specifies that an API key is required for authentication
        - userMaybeSignedIn: []
  /addressList/delete:
    post:
      operationId: deleteAddressLists
      summary: Delete address lists
      description: Deletes address lists. Must be created off-chain.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAddressListsRouteRequestBody'
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/iDeleteAddressListsRouteSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
      parameters:
        - name: x-api-key
          in: header
          description: BitBadges API Key for authentication
          required: true
          schema:
            type: string
      security:
        - apiKey: [] # This specifies that an API key is required for authentication
        - userIsOwner: []
        - userSignedIn: []
...
components:
  responses:
...
    BadRequestResponse:
      description: Bad Request - Invalid configuration or request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedResponse:
      description: Unauthorized - Does not meet authnetication requirements. This can be due to an invalid API key, CORS origin, or session cookie.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerErrorResponse:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Requires a BitBadges API key for requests
      x-type: string
    frontendOnly:
      type: apiKey
      in: header
      name: Origin
      description: Can only be called by the official BitBadges frontend. Not for developer use.
    userSignedIn:
      type: http
      scheme: cookie
      description: User must be signed in with a valid session.
    userMaybeSignedIn:
      type: http
      scheme: cookie
      description: User must be signed in for certain requests, but not all.
    userIsManager:
      type: http
      scheme: cookie
      description: User must be signed in and have manager permissions.
    userIsOwner:
      type: http
      scheme: cookie
      description: User must be signed in and have owner permissions.
  schemas:
    ErrorResponse:
      type: object
      required:
        - errorMessage
      properties:
        error:
          type: object
          description: Serialized error object for debugging purposes. Technical users can use this to debug issues.
        errorMessage:
          type: string
          description: UX-friendly error message that can be displayed to the user. Always present if error.
        unauthorized:
          type: boolean
          description: Authentication error. Present if the user is not authenticated.
    NumberType:
      anyOf:
        - type: string
        - type: number
      example: '1'
      description: String numbers are always returned to avoid precision loss. You may pass in request bodies as numbers or strings. See SDK docs for NumberType conversions.
...
    iListActivityDoc:
      properties:
        _docId:
          description: A unique stringified document ID
          type: string
        _id:
          description: A uniuqe document ID (Mongo DB ObjectID)
          type: string
        timestamp:
          description: The timestamp  of the activity.
          $ref: '#components/schemas/NumberType'
        block:
          description: The block number of the activity.
          $ref: '#components/schemas/NumberType'
        _notificationsHandled:
          description: >-
            Whether or not the notifications have been handled by the indexer or
            not.
          type: boolean
        listId:
          description: The list ID of the list.
          type: string
        addedToList:
          description: >-
            Whether or not the address is included in the list. Note that this
            could mean added to an whitelist or a blacklist
          type: boolean
        addresses:
          items:
            type: string
          description: The list of addresses that were added or removed from the list.
          type: array
        txHash:
          description: The transaction hash of the activity.
          type: string
      required:
        - _docId
        - timestamp
        - block
        - listId
      additionalProperties: false
      description: |-
        Fields for the MongoDB database document
      type: object
...
    iLatestBlockStatus:
      properties:
        height:
          description: The height of the latest block
          $ref: '#components/schemas/NumberType'
        txIndex:
          description: The transaction index of the latest block
          $ref: '#components/schemas/NumberType'
        timestamp:
          description: The timestamp of the latest block (milliseconds since epoch)
          $ref: '#components/schemas/NumberType'
      required:
        - height
        - txIndex
        - timestamp
      additionalProperties: false
      type: object
    iStatusDoc:
      properties:
        _docId:
          description: A unique stringified document ID
          type: string
        _id:
          description: A uniuqe document ID (Mongo DB ObjectID)
          type: string
        block:
          $ref: '#/components/schemas/iLatestBlockStatus'
          description: The latest synced block status (i.e. height, txIndex, timestamp)
        nextCollectionId:
          description: The next collection ID to be used
          $ref: '#components/schemas/NumberType'
        gasPrice:
          description: The current gas price based on the average of the lastXGasAmounts
          type: number
        lastXGasAmounts:
          items:
            $ref: '#/components/schemas/NumberType'
          description: The last X gas prices
          type: array
        lastXGasLimits:
          items:
            $ref: '#/components/schemas/NumberType'
          description: The last X gas limits
          type: array
      required:
        - _docId
        - block
        - nextCollectionId
        - gasPrice
        - lastXGasAmounts
        - lastXGasLimits
      additionalProperties: false
      description: |-
        Fields for the MongoDB database document
      type: object
...
    iMetadata:
      properties:
        name:
          description: The name of the badge or badge collection.
          type: string
        description:
          description: The description of the badge or badge collection.
          type: string
        image:
          description: The image of the badge or badge collection.
          type: string
        video:
          description: >-
            The video of the badge or badge collection. If a standard video is
            used, this should be a link to the video. We will use image as the
            poster image. If a youtube video is used, we embed it as an iframe.
          type: string
        creator:
          description: The creator of the badge or badge collection.
          type: string
        color:
          description: The color of the badge or badge collection.
          type: string
        category:
          description: >-
            The category of the badge or badge collection (e.g. "Education",
            "Attendance").
          type: string
        externalUrl:
          description: The external URL of the badge or badge collection.
          type: string
        tags:
          items:
            type: string
          description: The tags of the badge or badge collection
          type: array
        socials:
          additionalProperties:
            type: string
          description: The socials of the badge or badge collection
          type: object
        offChainTransferabilityInfo:
          properties:
            host:
              type: string
            assignMethod:
              type: string
          required:
            - host
            - assignMethod
          additionalProperties: false
          description: The off-chain transferability info of the badge or badge collection
          type: object
        attributes:
          items:
            properties:
              type:
                enum:
                  - date
                  - url
                type: string
              name:
                type: string
              value:
                anyOf:
                  - type: string
                  - type: number
                  - type: boolean
            required:
              - name
              - value
            additionalProperties: false
            type: object
          description: The attributes of the badge or badge collection
          type: array
        fetchedAtBlock:
          description: The block the metadata was fetched at.
          $ref: '#components/schemas/NumberType'
        fetchedAt:
          description: The time the metadata was fetched.
          $ref: '#components/schemas/NumberType'
        _isUpdating:
          description: Whether the metadata is currently being updated.
          type: boolean
      required:
        - name
        - description
        - image
      additionalProperties: false
      type: object
    iBitBadgesAddressList:
      properties:
        listId:
          description: The ID of the address list.
          type: string
        addresses:
          items:
            type: string
          description: The addresses of the address list.
          type: array
        whitelist:
          description: >-
            Whether or not to include ONLY the addresses or include all EXCEPT
            the addresses.
          type: boolean
        uri:
          description: The URI where to fetch the address list metadata from.
          type: string
        customData:
          description: Arbitrary custom data that can be stored on-chain.
          type: string
        createdBy:
          description: The address that created the address list.
          type: string
        aliasAddress:
          description: The alias cosmos address of the address list.
          type: string
        _docId:
          description: A unique stringified document ID
          type: string
        _id:
          description: A uniuqe document ID (Mongo DB ObjectID)
          type: string
        updateHistory:
          items:
            properties:
              txHash:
                type: string
              block:
                $ref: '#components/schemas/NumberType'
              blockTimestamp:
                $ref: '#components/schemas/NumberType'
            required:
              - txHash
              - block
              - blockTimestamp
            additionalProperties: false
            type: object
          description: The update history of this list
          type: array
        createdBlock:
          description: The block number when this list was created
          $ref: '#components/schemas/NumberType'
        lastUpdated:
          description: >-
            The timestamp of when this list was last updated (milliseconds since
            epoch)
          $ref: '#components/schemas/NumberType'
        nsfw:
          properties:
            reason:
              type: string
          required:
            - reason
          additionalProperties: false
          description: The NSFW reason if this list is NSFW
          type: object
        reported:
          properties:
            reason:
              type: string
          required:
            - reason
          additionalProperties: false
          description: The reported reason if this list is reported
          type: object
        private:
          description: True if this list is private and will not show up in search results
          type: boolean
        viewableWithLink:
          description: True if this list is viewable if queried by the list ID directly
          type: boolean
        metadata:
          $ref: '#/components/schemas/iMetadata'
          description: The metadata of the address list.
        listsActivity:
          items:
            $ref: '#/components/schemas/iListActivityDoc'
          description: The activity of the address list.
          type: array
        views:
          additionalProperties:
            properties:
              ids:
                items:
                  type: string
                type: array
              type:
                type: string
              pagination:
                $ref: '#/components/schemas/PaginationInfo'
            required:
              - ids
              - type
              - pagination
            additionalProperties: false
            type: object
          description: The views of the address list.
          type: object
        editClaims:
          items:
            properties:
              claimId:
                type: string
              plugins:
                items: {}
                type: array
            required:
              - claimId
              - plugins
            additionalProperties: false
            type: object
          type: array
      required:
        - listId
        - addresses
        - whitelist
        - uri
        - customData
        - _docId
        - updateHistory
        - createdBlock
        - lastUpdated
        - listsActivity
        - views
        - editClaims
      additionalProperties: false
      description: |-
        Fields for the MongoDB database document
      type: object
    GetAddressListsRouteRequestBody:
      properties:
        listsToFetch:
          items:
            properties:
              listId:
                type: string
              viewsToFetch:
                items:
                  properties:
                    viewId:
                      type: string
                    viewType:
                      enum:
                        - listActivity
                      type: string
                    bookmark:
                      type: string
                  required:
                    - viewId
                    - viewType
                    - bookmark
                  additionalProperties: false
                  type: object
                type: array
              fetchPrivateParams:
                type: boolean
            required:
              - listId
            additionalProperties: false
            type: object
          description: >-
            The lists and accompanyin details to fetch. Supports on-chain,
            off-chain, and reserved lists.
          type: array
      required:
        - listsToFetch
      additionalProperties: false
      type: object
    iGetAddressListsRouteSuccessResponse:
      properties:
        addressLists:
          items:
            $ref: '#/components/schemas/iBitBadgesAddressList'
          type: array
      required:
        - addressLists
      additionalProperties: false
      type: object
    UpdateAddressListsRouteRequestBody:
      properties:
        addressLists:
          items:
            allOf:
              - $ref: '#/components/schemas/iAddressList'
              - properties:
                  private:
                    type: boolean
                  editClaims:
                    items:
                      properties:
                        claimId:
                          type: string
                        plugins:
                          items: {}
                          type: array
                      required:
                        - claimId
                        - plugins
                      additionalProperties: false
                      type: object
                    type: array
                  viewableWithLink:
                    type: boolean
                required:
                  - editClaims
                additionalProperties: false
                type: object
          description: |-
            New address lists to update.
            Requester must be creator of the lists.
            Only applicable to off-chain balances.
          type: array
      required:
        - addressLists
      additionalProperties: false
      type: object
    iUpdateAddressListsRouteSuccessResponse:
      additionalProperties: false
      type: object
    DeleteAddressListsRouteRequestBody:
      properties:
        listIds:
          items:
            type: string
          description: The list IDs to delete.
          type: array
      required:
        - listIds
      additionalProperties: false
      type: object
    iDeleteAddressListsRouteSuccessResponse:
      additionalProperties: false
      type: object
...
    PaginationInfo:
      properties:
        bookmark:
          type: string
        hasMore:
          type: boolean
      required:
        - bookmark
        - hasMore
      additionalProperties: false
      description: |-
        Type for pagination information.
      type: object
...
    iGetStatusRouteSuccessResponse:
      properties:
        status:
          $ref: '#/components/schemas/iStatusDoc'
          description: Includes status details about the indexer / blockchain.
      required:
        - status
      additionalProperties: false
      type: object
...
    iAddressList:
      properties:
        listId:
          description: The ID of the address list.
          type: string
        addresses:
          items:
            type: string
          description: The addresses of the address list.
          type: array
        whitelist:
          description: >-
            Whether or not to include ONLY the addresses or include all EXCEPT
            the addresses.
          type: boolean
        uri:
          description: The URI where to fetch the address list metadata from.
          type: string
        customData:
          description: Arbitrary custom data that can be stored on-chain.
          type: string
        createdBy:
          description: The address that created the address list.
          type: string
        aliasAddress:
          description: The alias cosmos address of the address list.
          type: string
      required:
        - listId
        - addresses
        - whitelist
        - uri
        - customData
      additionalProperties: false
      type: object
...